drivers: sensor: iim4623x: basic sensor support
Add support for the TDK Invensense IIM-46230 and IIM-46234 motion
tracking devices by implementing the most basic sensor sample_fetch and
channel_get API.

The iim4623x devices can be connected through UART or SPI, but this commit
solely deals with the SPI connection.
The two sensors are has the same protocol and contain almost identical
feature sets. The biggest difference seems to be difference in supported
full-scale resolution on both the gyroscope and accelerometer.

The interface to the iim4623x devices is a bespoke protocol which at its
core is command based though a register map is implemented on top. Given
the command based nature of the protocol, physical reads and writes must be
sequenced according to the chip status which is signalled using a DRDY gpio
(data-ready). As an example, to read a register the logical flow must
follow:
    1. Send "read register" command
    2. Await DRDY gpio assertion from the iim4623x
    3. Read the response packet

Most configuration of the iim4623x devices is performed through the
register map, but other commands exist. For instance:
    * Read firmware version
    * Start/stop streaming data
    * Select streaming interface
    * Set UTC time
and more.
Almost all commands will receive a response packet, the ones that don't are
typically tied to the device streaming mode.

To support this protocol, the device driver adds a generic helper function
to build protocol packets. Using this, functions for reading/writing
registers are then implemented.
These helpers take care of boilerplate actions such as checksum validation.

In order to implement the protocol operations as synchronous logical
operations, the driver adds an iim4623x_bus implementation which utilises
the RTIO subsystem to create chains of operations.
It is worth noting that the driver private data keeps an rtio_sqe pointer
during transactions in order for the rtio execution to await a completion
signal when the data-ready interrupt asserts.
It is also worth noting that the device datasheet specifies a minimum time
to wait between protocol operations. This is currently added at the end of
the operation chains.

As for configuration, the output data rate (odr), full-scale values for
the accelerometer and gyroscope, bandwidth configuration for the
accelerometer and gyroscope, can be specified in the devicetree in a
similar fashion to that of similar device drivers (icm45686 for instance).

While the driver does not yet support the trigger and stream APIs, the
driver is laid out to accomodate these APIs after some additional
development.

Signed-off-by: Emil Dahl Juhl <emil@s16s.ai>