drivers: sensor: sch16t: read and decode driver
This adds the initial driver for the Murata SCH16T series of combined
accelerometer and gyroscope sensors. The driver is based on the read and
decode API since the fetch and get API is planned to be deprecated.

The SCH16T can be interfaced with using a classic register read/write
protocol. The SPI frames are constructed according to the SafeSPI 2.0
specification. As a result, the device supports both 48-bit and 32-bit SPI
frames. In order to reduce complexity, only 32-bit frames are supported in
this driver. No features should be lost by opting to use 32-bit instead of
48-bit.

The SCH16T supports multi-slave operation on a single bus by use of the
DRY_SYNC pin. Since this pin is also used for signaling interrupts for data
ready, it is configured as such and multi-slave support on a single bus is
not supported by this driver.

Despite this, support for specifying values for decimation of the readings
and respective dynamic range was implemented using device tree bindings.

Signed-off-by: Pedro André <pedro@sentrytechnologies.ai>