drivers: sensor: rm3100: use edge based interrupts
The rm3100 streaming logic currently relies on using its interrupt pin as a
level based interrupt. Edge based interrupts can also be used provided some
initial polling of the pin is done to check if it is active already.
Considering that hardware support for edge based interrupts is more
frequent than level based interrupts, using edge is desirable as it makes
the driver compatible with a wider range of devices. As a result, convert
the current logic to use edge based interrupts.

Signed-off-by: Pedro André <pedro@sentrytechnologies.ai>