drivers: gnss: f9p: add option to perform reset on init
The U-Blox F9P module saves the current configuration dynamically, that is,
if the module is configured to emit a certain message every second, for
example, that setting will be saved and committed to non-volatile memory
immediately. Such is desirable such that the device does not need to be
configured constantly. However, given the current way the zephyr driver
manages the device, it is configured on initialization anyway. In order to
ensure that the device always starts from factory settings, add a device
tree binding that allows reseting the device as the first step of
initialization.

This feature also has another advantage: the F9P module stores satellite
information in volatile memory that can be kept alive for up to two weeks
due to an on-board battery on the module. Such is not ideal when it is
necessary to verify if any configuration change makes a difference. Since
the reset binding also clears this volatile memory, it is an ideal option
to use while developing.

Signed-off-by: Pedro André <pedro@sentrytechnologies.ai>