gnss: rtk: fix multiple instances of same device
When multiple instances of the same type of RTK enabled GNSS device are
instantiated on the same board, there will be a collision with their RTK
data callbacks. The u-blox F9P module is a clear example of that, where the
following compilation error is reported:

     error: redefinition of '_gnss_rtk_data_callback__f9p_rtk_data_cb'

Similar to the solution used at d36a0fcb950 (drivers: gnss: allow same
callback), use `__COUNTER__` instead of `_callback` in the generated entry
to avoid collisions.

Signed-off-by: Pedro André <pedro@sentrytechnologies.ai>