drivers: gnss: f9p: rtk: add station mode support
This adds station mode support to the F9P GNSS driver. The current approach
currently consists in using a DTS ENUM binding to determine which RTK mode
the device should be configured as:

    - none
    - station
    - rover

Obviously this approach is not ideal as it is F9P specific and also static.
A probably better solution would be to add subsystem support to configure
the device as station or rover from userspace with all the extra config
necessary. For now, this solution is enough to validate the concept for
static configurations. Despite this, in the future, we should work on a
subsystem level approach.

For the station specific configuration handled in this, the device is
configured to output all relevant RTCM messages on UART2. Furthermore, it
is instructed to perform a survey in with the minimum duration and accuracy
limit specified in the dts. Once the survey in finishes, the station device
changes its GNSS fix to "time" and it starts emitting RTCM correction data.

Signed-off-by: Pedro André <pedro@sentrytechnologies.ai>