drivers: gnss: f9p: rtk: rover: get relposned
As soon as a U-Blox module starts receiving RTCM correction data, it will
automatically transition into `rover` mode. This means that the navigation
data it outputs will be several degrees of magnitude more precise, being
even able to achieve sub centimeter precision. Despite this, to obtain a
the static heading between a specific rover and the station emitting
correction data to it, it is necessary to subscribe to a new type of U-Blox
message: UBX_NAV_RELPOSNED.

Ideally, every-time this message was received, a callback could be
registered just like what is done for GNSS data, satellite data and RTK
data. This will probably be the right solution moving forward, but since it
is a U-Blox specific feature, to simplify for now, the static heading is
injected into the normal GNSS data if the received static heading is valid.

Signed-off-by: Pedro André <pedro@sentrytechnologies.ai>